Peiority Bugfix, working on RV Mode...
Past update had broken a few things in bad ways...
Otherwise, have made some functional tweaks to RV Mode in an attempt to try to be able to bridge compatibility with Linux RV binaries...

+++ tk_qsrc/tk_clib/vxcore/include/tk_tkdfs.h
@@ -166,7 +166,7 @@ byte ln2_inosz;
 TKDFS_InodeInfo img_inocache[64];
 
 int img_blkcache_blkid[128];
-int img_blkcache_data[128*512];
+byte img_blkcache_data[128*512];
 int img_blkcache_flag[128];
 
 u64 pad0;	//alignment pad
